Initial modular gunsprite update (#204)
* Initial commits
Add base framework, example CZ97, example FNX
FNX is able to be reskinned to one of four combinations, built right in

* Fixes issues present in initial commit, adds reskin ability to guns

* Adds the CZ97 and FNX example guns to sec vendor
The security pistol which I replaced the sprites on at one point is in the vendor. Let's add these two in as well, so they can at the very least get some use.

* Adds the CZ97 and FNX example guns to sec vendor
The security pistol which I replaced the sprites on at one point is in the vendor. Let's add these two in as well, so they can at the very least get some use.
Co-authored-by: BaronVonTerra <zswatzel@gmail.com>

* Create initial documentation.