Background noise reduction (#208)
* Gives ATC its own channel

* Separates ATC channels
Shuttles now operate on the local ATC channel, whereas all other traffic operates on the regional channels.
Also fixes a mislabel of an eclipse edit as a "Mithra edit". I don't even. I just cannot even. I am unable to even right now.

* Headsets.
Whoops.

* Maps in ATC radios for the hangars
Maps in some local ATC radios for the hangars, so players can hear when a shuttle takes off or lands when their headset has the channel disabled. This also applies to the outpost docking pads.

* NO CHANGES - MESSAGE ONLY
Halting work temporarily - need to make a config file for the juicy config options instead of having everything set at compile

* NO CHANGES - MESSAGE ONLY
The changes that I want to finalise require the config options PR to be merged, but outside of that, this is perfectly functional and ready for review.