Patch for antag selection (#272)
* Adds more checks for antag/job selection incompatibilities

* Adds more checks for antag restrictions.
For some reason, the job controller has priority over the antag controller. This should not be the case, but I'm still figuring out how to get around that. We WANT the antag controller to have priority, because we want people who are rolling for antag but have like Security on high to be able to get antag every once in a while without changing off of security.

* just gonna push this for PR
gonna let Nestor decide whether this is in compliance with the freeze or whether it counts as a feature addition.
Quite honestly this is kicking my ass and I'm... beat down with all the fucking mapping oversights caused by my stupidity in porting it. I'm not feeling very good right now... may as well push what i have which will patch the issue but not outright fix it.